스파인 MeshAttachment_clone, WeightMeshAttachment_clone 함수 구현 중, triangles 복제 코드 메모리 오버플로 버그 수정

+++ spine-c/src/spine/Attachment.c
@@ -220,7 +220,7 @@ spMeshAttachment* spMeshAttachment_clone( spMeshAttachment* source )
   if ( source->triangles != NULL )
   {
     self->triangles = MALLOC( unsigned short, source->trianglesCount );
-    memcpy( self->triangles, source->triangles, ( sizeof( int ) * source->trianglesCount ) );
+    memcpy( self->triangles, source->triangles, ( sizeof( unsigned short ) * source->trianglesCount ) );
   }
   
   // edges, edgesCount
@@ -298,7 +298,7 @@ spWeightedMeshAttachment* spWeightedMeshAttachment_clone( spWeightedMeshAttachme
   if ( source->triangles != NULL )
   {
     self->triangles = MALLOC( unsigned short, source->trianglesCount );
-    memcpy( self->triangles, source->triangles, ( sizeof( int ) * source->trianglesCount ) );
+    memcpy( self->triangles, source->triangles, ( sizeof( unsigned short ) * source->trianglesCount ) );
   }
   
   // regionUVs, uvs, uvsCount
